CMake configuration option: WITH_PIC
Introduces a new onfiguration option, WITH_PIC, which can be
used to force compile all build objects as position independent
sources. The motivation behind this option is to provide a means
for the end user to manually control the way libwiredtiger is
compiled. If building a static version of wiredtiger, WITH_PIC
allows the end user to configure whether the sources should be
position independent, based on their needs for linking the
wiredtiger artifact.